our backbone object creation was broken, since we auto generate IDs, backbone does not know what is a new model and what needs to be created on the server, rather than updated.
we use a attribute, created, to tell us whether something is created.  it is false by default.  on a successful creation, it is set to true.
we override collections.create and model.isNew for this
we also modify every model class to update model defaults rather than defining new defaults this is because we need the inherited docs property to propagate through for at model creation