Update README.md
Corrected definition of bokeh. "Bokeh" refers only to the aesthetic quality of the blurred background; the technique of blurring the background in order to direct attention to the foreground goes under the more prosaic terms of "selective focus" or "shallow focus." See the Wikipedia page or Mike Johnston's description here (http://theonlinephotographer.typepad.com/the_online_photographer/2009/01/what-is-bokeh.html).