Moving DataSpec and ColorSpec into properties.py, and merging functionality of MetaGlyph into MetaHasProps, and removing MetaGlyph
The root purpose of this was so that I could put ColorSpecs into the mix-in classes of LineProps and FillProps.

TextProps does not declare its color to be a ColorSpec, because in general, text properties are commonly used in places where color is not data-driven.  If data-varying color is desired for text, then explicitly redeclare the relevant text properties as ColorSpecs in the Glyph subclass itself.