running the bokehserver with splitjs False and devjs True does not work.
This is because there is an inconsistency with build/release js paths, and
what is on the server.

I think it is easier to standardize on our bokeh.js living inside build/js, or release/js

This commit
- modifies the Gruntfile to build into build/js and release/js
- modifies setup.py to look for files in that directory
- renames JSROOT to BOKEHJSROOT, since the name JSROOT is a bit misleading, as we have things like css in there
- fixes an os.path.join call which had 'static/js' as an arguemnt, since that's not platform agnostic

This isn't really tested yet, so I would merge this after I get a chance to test it but I wanted to open the PR first.