Set {np.}random.seed(1) in examples/test
This allows us to run examples as tests and always get the
same plots, even if random data is involved. This solution
doesn't require any changes to examples. The slight drawback
is that tracebacks contain additional <string> entry at the
beginning, thought the remaining entries and file names are
preserved.