fix "python setup.py build"
When running "build" the --install_js option got added, but in parse_jsargs an error got thrown, if not in 'install', 'develop', 'sdist', 'egg_info' mode. So build always produced an error. (ran across this while packaging bokey as rpm where the build and install are two separate steps).

Hope this fixes it, but perhaps there is a better way.