Wrap entire code in a null define
This is required to keep Backbone from seeing Require.js' `define`
inside IPython Notebook and trying to use that instead of returning the
proper values.  Tried a few different methods and this is the one that
seems to work.

Note, testing in IPython Notebook requires clearing all the cells *and*
restarting the notebook server to ensure things are working as expected.