Add a simple "callback" action
Accepts a dictionary that maps plot objects to names that will be
available to the JS code, as well as a snippet of JS code that serves
as the function body. The function also receives a "value" parameter,
which is, for instance, the new value of a widget when it is updated.