Add CSV endpoint for ColumnDataSources
Given that there is a very simple way to generate CSV files from pandas DataFrames, I've added an endpoint that outputs a CSV file representing a ColumnDataSource. 

If the format is unrecognized or the type is not ColumnDataSource when the format is CSV, this function defaults to returning the JSON.

Maybe it would be nice to have a plot Tool that provides a link to the DataSource's CSV endpoint. (Like "Export to CSV...").