Link updates with tab widget's active attribute
Closes #2455

When a Tabs widget's active attribute is updated, the corresponding
html should be re-rendered. Conversely, when a tab link is clicked,
the active attribute of the python widget object should be updated.

Previously, the only way I could find to update tabs was to create
a new Tabs python widget object, which however seemed to lead to
an assortment of bugs with plot renderings.