PreviewSaveTool init doesn't create dom element
(Minor) Right now every time a figure is initialized with the
PreviewSaveTool, a corresponding dialog div element gets created.
Pretty minor since the div elements aren't visible, but it is rather
confusing since it makes it look like updating figures in python is
a wrong thing to do. For example, try starting stock_app after
adding the "save" tool to each of the plots, changing the plot sources,
and looking at a html console.

This still doesn't entirely fix the problem in that after a figure
is removed, the corresponding PreviewSaveTool dom element nor view
don't seem to get removed. My guess is that its because a reference
to the PreviewSaveTool model isn't kept by the app (if we are creating
the tool via the "save" string). If that's the case maybe Plot's
remove() method can be overloaded to call all of its tools' remove()
methods.

Update some docs to change SaveTool to PreviewSaveTool, and dimension
to dimensions.