Plot view removal also remove the tool views
When a Plot backbone view is remove()'d, it will also call remove()
on its tools. Otherwise right now the tools do not seem to be included
with other models and so won't be remove()'d when their plot is removed.
I think this should be ok as long as its ok to call remove() twice on a
view.