Add spin counter to Icon
Add a spin counter, which a bokeh-server script can listen to see
if the icon's spin state has updated. A python script will then be
able to update the icon's spin state, receive a notification of the
spin update, and then perform some other actions, as in the added
example.