Do not scroll datatable if selection already in viewport
If any of the selection rows are already in the viewport range,
do not scroll the data table. This way selections on the data
table itself will not cause re-scrolls. Additionally, offer a
boolean parameter for disabling this scrolling to allow the
old behavior to be retained. (But why would you want to. =P)