Add Component base class above Plot and Widget
Previously Plot inherited from Widget, making it technically a widget,
which is not. Often the documentation referred to Widget as Plot/Widget
pair, to make it more prominent that widgets include Plot as well. With
bokehjs split this was even worse, because we couldn't use Widget base
class to determine if we need bokeh-widgets.js or not.

Thus Component replaces Widget and new Widget abstract base class, that
inherits from Component, is now located in models/widgets/widget.py. If
something inherits from Widget, it is a widget, and is optional from bokeh
being a pure plotting library perspective. Plot inherits from Component.
Now you don't have to refer to Plot/Widget pair, just say Component.