Remove message hmac for the time being
This should be a property of the transport not the message, so that
only transports that need it would have it.

TCP should not need an HMAC because TCP is a reliable protocol and
establishes a persistent connection, so we could do authentication
once at the start of the TCP connection and not per-message.