Create a separate websocket route for each Application on the server
This simplifies a lot of things; without it, each message on the websocket
would potentially have to specify the application name, and some public APIs
that take a session id might need to take an app name too.

In a production deployment, there's no value I can think of to putting all applications
from the same browser tab on the same cluster node. In fact it might be better to separate
them. Multiple apps from one browser tab should be a weird corner case anyhow.

We only support multiple apps as a kind of efficiency hack (multiple apps deployed
in a single process) so a good guideline might be to avoid leaking the fact
of multiple apps into the API, except as part of the server URL. That's what
this commit does - the app is now specified by the URL you provide to ClientConnection.
This also allows people to transparently change their deployment scenario to use
separate nodes for each app, if they want, if they have a reverse proxy that
can route by the path prefix.