Include underscore and coffee-script only once in package.json
Both are dev and runtime dependencies, so originally were included
in devDependencies and dependencies sections of package.json. This
causes npm to raise annoying warnings. This is because, whether a
dependency is dev or not, it will sill land in node_modules/, so
technically there's a collision. Lets silence this.