Recursively serialize JS HasProperties, but exclude models the server doesn't know about
The previous code didn't recurse into property values, so missed some models;
but when adding recursion, there are things in attributes that should not
be serialized. Also there are whole classes such as LayoutBox that should
never be visible to Document. So this commit tries to deal with that; it
could probably be done more cleanly by refactoring HasProperties into
a separate types for "model that goes in a Document" vs. "Backbone.Model" vs.
"thing with computed Properties"