Expire server sessions after a timeout rather than immediately
This is needed if output_server is going to output to a session,
because we don't want the session to instantly vanish. We may want
it to never vanish ... but that will require more work, since a production
server does need to expire these things or it will run out of memory.