Dedicate each websocket to a single session.
The reason for this is that for a clustered server, each session
would live canonically on one cluster node; so a websocket
may not be on the right node to interact with a given session.
A router layer would have to route each socket to the appropriate
node for that session.

If we did "user affinity" instead of "session affinity" and assumed
(probably a good idea) that we never share sessions between users,
we could do multiple sessions per socket. So maybe this commit won't
turn out to be ideal, but for now 1-socket-1-session seems like a
good guess.

The client APIs now keep the connection internal, so
potentially we could transparently share connections among sessions
in the future - the new JS/Python APIs with only ClientSession public
reveal a bit less about the implementation.