Unify accumulate_from_superclasses caching, and avoid hasattr() for the cache
Also, use the cache for class_properties().

The use of hasattr() was broken because if we cached for a base class
first and then for a subclass, we would only have the base class props
for the subclass. This problem was mostly hidden because we did not
cache in class_properties() (now we do), and because typically we cached
for the subclass first.