Make socket writes async and always lock ServerSession.document
These two are entangled because both the lock and the async writes
cause coroutine behavior to "spread" further up the stack.

The main benefit of async writes is to limit the write backlog. With
fire-and-forget writes, if our outgoing buffer to the client is full,
we just keep creating more and more pending write tasks. With this
patch to wait for each write to finish before we write the next thing,
that shouldn't happen.

A new issue is that we'll hold the session lock until we can finish
writing out all the events we generated by changing the document. This
means that a client that never reads events could "lock up" a session;
but shared sessions aren't secure anyway so it probably doesn't matter.

This patch takes the session lock during timeouts and periodics, which
we didn't do before. This should keep us from writing out messages
from a timeout/periodic callback in the middle of doing something else.