Do not default NumberSpec props to field name
In Bokeh 0.10 and now, the CoffeeScript default was undefined
(or 0, for some angles). After the tornado merge, we were sending
the field name default from Python when we did not ever use that
name before. In 0.10, we'd never send the default so CoffeeScript
never saw it.

The field name doesn't seem like a good default because in general
it ends up being very magical and unclear; if I type
`Line(source=source)` then it magically uses the `x` and `y` columns,
but if I later rename those columns, or if I'm adapting an example
and my columns aren't named that, it gets confusing what's going on.