charts.glyphs, do not set properties that weren't provided
If we want a property to be the default, we need to omit
it from kwargs, not set it to the default value in kwargs.

This was a conceptual problem (was the property set or
was it the default?) but it's also a practical problem
because HasProps shouldn't have to handle a getter
before its __init__ has been called.