Builder: convert some class attrs into plain class attrs not Property
Property is not a way to specify class attributes. A Property when
provided as a class attribute defines an instance attribute with the
same name. In these cases, it looks like we were providing the Property
in order to define an "abstract class attribute" which subclasses
were supposed to replace by a plain non-Property class attribute, which
is not what Property does, and having the Property in base class essentially
did nothing, as far as I know...

The immediate reason to change this is that I'm working on a modification
to HasProps where it's no longer OK to get a property's value prior to
HasProps.__init__, which Builder needed to do when these things were
properties - but I guess they shouldn't have been properties anyway?

I'm not sure I fixed all of these, for now these are the ones which
were used prior to calling HasProps.__init__ which created an immediate
issue. Someone who knows Builder better might be able to de-Property
a few more of them.