Split Property into Property and PropertyDescriptor
The Property is the binding to a particular name and class; the
PropertyDescriptor is the generic type, default, and help information.

This forces cleaner code factoring so that we can have type parameters
and _units_type that we're confident won't try to mess with any
HasProps instance attributes.

In this commit, PropertyDescriptor.transform() somewhat oddly takes
an object instance and a name; this is intended to set the stage for
merging to_dict() into transform(), but it may be that we can simply
implement the foo_units properties in a better way now by having
a separate subtype of Property that chains to an underlying Property.
In that case we don't need transform() to be weird. Anyhow, enough
stuff in this commit so will continue to evolve things.