Add properties.Override allowing changes to defaults in subclasses
Also use this to clean up some of the warnings about bad overrides.

Most/all of the remaining override warnings are because of attempts
to override the type, which are conceptually broken, not something
we can fix with new syntax. Most of these need to drop the property
from the base class and just have it in the subclasses.