Handle property defaults that are unstable by saving the value as a non-default
An unstable default is one that's different every time we compute the default,
or that could become different behind our backs or per-HasProps-instance.

We were handling the list/dict container case of this, but not the cases of Model
or of a default computed from a fucntion. In these cases we'll have to always act
as if the app has set the value, so it will never appear to just be the default...
we could in theory avoid this for Model by watching for changes to the model, though.