If an app directory contains theme.yaml, set it as document.theme
This allows apps to ship with a custom theme.yaml.

As of this commit, however, this will break apps because it discards
default.yaml. The solution is probably to move default.yaml into
Python code, but if that fails, we could allow themes to be based
on other themes or something.