Store ScriptHandler module on the Document not on ScriptHandler
This is intended to fix #3275, which looks like:

 ERROR:bokeh.server.protocol.server_handler:error handling message Message 'PATCH-DOC' (revision 1): AttributeError("'NoneType' object has no attribute 'value'",)

The problem is that a global in the sliders.py script has become None,
so when we try to mess with amplitude.value it produces the above error.

Theory: when a module is GC'd, all its globals are set to None.

If so, the module would be gc'd if we have >1 session because we were
storing the module on the ScriptHandler instance and overwriting it
for every modify_document. But with only 1, the field on ScriptHandler
prevented GC.

Storing the most recent module in ScriptHandler makes no sense anyhow,
GC or no GC.

Fixed to store it on Document instead which will prevent
GC as long as the document exists. This should work out pretty well in
practice though in theory it could keep the module around longer than
necessary if someone was doing something odd.