Add signed session IDs and factor session ID generator into utils
In a future commit, the session ID generator will be used on the
client side, so moving it to utils. This also makes it easier to
test.

The signed session ID feature allows restricting session creation
to processes that have the same BOKEH_SECRET_KEY as the server.

So for example if a Flask/Django/whatever app wants to authorize
all new Bokeh sessions, it can share a BOKEH_SECRET_KEY with the
Bokeh server, and call bokeh.util.session_id.generate_session_id()
to generate valid session IDs. Without the secret key, it will be
impossible for anyone else to create valid IDs.

A Bokeh server running by itself can also require signed session IDs,
which would mean that the only way to get a session is to go through
doc_handler.py (load the app as a web page), which generates a new
session.

We may want an additional setting for the Flask/Django/etc. case where
doc_handler.py will NOT generate a fresh session ID and will require
one to be passed in. In this commit we don't have that, so the
Flask/Django/etc. process can't mediate all session creation, but
we'll add it soon.