When we customize property serialization, also customize deserialization
This fixes a problem with examples/plotting/server/glyphs.py where
we'd have AnnularWedge.inner_radius as a single value, we'd send it
to the client as a dict, at some point the client sent it back to us
as a dict, we then stripped the units out and emitted a ModelChangedEvent
as a dict sans units (because ModelChangedEvent contains the new regular
value, not the new serializable_value()). This new units-less dict was
then sent to the client causing the client to use default units instead
of the originally-provided units.

To fix this we need the serializable_value() in ModelChangedEvent.

We also need to be able to get a change event back from the client
without "forgetting" the dict-ness of our local attribute.
"dict-ness" means that we remember whether you set `foo = 42` or
`foo = { 'value' : 42 }`.

So we move the logic that does from_json and then sets the prop
from Document into a HasProps.set_from_json which in turn delegates
to Property, so that DataSpecProperty can process the JSON and
decide to preserve its dict-ness in the JSON case but not in the
regular `__set__` case.