Secret key only needs 32 bytes (256 bits) of entropy for SHA-256 HMAC
Of course, looking at a string of chars we don't really know how many
bytes of entropy it has... probably a 32-char string has less than 32
bytes since we can't use all 0-255 values for each char. But things
don't entirely break if the entropy is too little, and we don't want
to warn about a key that could be acceptable. Less than 32 chars
is almost certainly not OK so do warn about that though.