Round-trip 'subtype' part of ref through CoffeeScript
We had `'subtype' in obj` instead of `'subtype' of obj` so
Coffee was just ignoring the subtype; which is good, because
it's supposed to, the intended code was bogus.

However, Coffee was also ending up creating refs without
the subtype in them, which made us generate a ton of bogus
change notifications during initialize() because the ref
with and without the subtype would not compare equal.

So this patch makes us keep the subtype as we round-trip
through coffeescript; that means Python will create the
right model on the server, and it means we don't send
all these bogus change events on startup.