add include_defaults=false mode on the CoffeeScript side and use it
This fixes our json diff when we load a Document; since the server
no longer sent defaults, we would diff the JSON we loaded vs.
what the server sent, and come up with all of the client-side
defaults as attributes the server was missing. We'd then send
all the defaults back to the server. Oops.