Fix tracking of _all_models with list-of-instance properties
The issue here was that HasProperties._value_record_references
ignored its 'recurse' flag which led to bogus behavior when
_tell_document_about_change tried to do its attach_document
and detach_document.

This fixes #3361, where for Plot.renderers in particular, we
thought the new Span was already in the document, because
_value_record_references was messed up and recursed too far.

In the long run, a better fix would be to port the way we do
_all_models in Python (constantly reconstructing it from scratch),
because the way we do it in Coffee cannot avoid leaks when there
are cycles. However, with this fix we avoid failing to include
models in _all_models that should be there. In memory management
terms we're replacing an early free (memory corruption) with
a leak.

The test case added here triggers the bug which was fixed, and
illustrates it in a simpler scenario than #3361.