Don't draw axis label when rule coords are NaN
This happens when there is no data, so ranges don't work and
thus we can't compute coords. I think we should be able to
draw axis label despite this, but for now at least don't draw
it in some random location (which appears as if there was some
garbage painted on canvas, e.g. misplaced ticks).