Remove axis.location property
We're going to start handling axes and annotations in a similar manner so that
annotations can be added to side panels. Legend, span, and axis have a location
property that do different things.

Location on span and legend are used. Where as axis.location is only used in
the most trivial of ways when setting up the layout renderers.  In addition
we don't have any examples or docs that demonstrate the use of this property
(probably becuase it doesn't do anything useful).

We could do a deprecation cycle on the python property and then fully
ignore it on the bokehjs side, but that seemed a little pointless.