Remove add_annotation method; never use p.renderers.extend
Fixes: #4267

* Remove add_annotation method in favor of add_layout (see below)
* Always use add_layout in examples & docs (never p.renderers.extend)

The add_annotation method is essentially the same as the add_layout
method and it seems unecessary (and a little confusing, at least to me) to have both.

The method has never been in a release of Bokeh so doesn't need to be deprecated.

We were already adding legends (which are annotations) with the add_layout
method, so I believe this is consistent. FWIW I don't think add_layout is the
most intuitive name, but it's the one we've got.