Split SidePanel code out of Axis
We want to reuse the same SidePanel behavior for annotations, so split it out
of the Axis code.

Plot calls add_panel for any renderer in side lists and then the panel sets
itself up. The side panel's update_constraints method is now outside of a class which
is a little odd, but can't think of another way to make it re-usable for
annotations too which don't share the same class hierarchy as Axis.

In addition, clean-up a little the use of _doc_attached so only put things in
_doc_attached if they need the document attached to the model (this cleans up
it's use to just a couple of models). In tests, don't call _doc_attached, but
use the attach_document method.