Make a new document with only one root
This goes all the way and makes a new document with only one root - that
of the model we're trying to show. If leaves the model on the main
global document though, which i believe is what's enabling push_notebook
to still work. Although that seems like magic to me!