Fix toolbar background-color issue; fix top border.
This fix highlights that we have a bug in layout with the edge of the
frame overlapping (or perhaps underlapping is a better word) with things
surrounding it. The 1px off meant that the border wasnt showing when
there was title annotation above. It also is now missing on the right,
being overlapped by the toolbar.