In a notebook div send a document with only the model needed for show (#4563)
* Only show the root_ids we care about

* Make a new document with only one root

This goes all the way and makes a new document with only one root - that
of the model we're trying to show. If leaves the model on the main
global document though, which i believe is what's enabling push_notebook
to still work. Although that seems like magic to me!

* Use a context manager and restore the original document

* Add test to check notebook_div uses new context manager

* Add tests for context manager