util/zoom::scale_range allows v_axis_only && h_axis_only
remove a TODO about possibly checking for v_axis_only && h_axis_only in util/zoom::scale_range.
The logic inside scale_range() will work (doing nothing) if both are true.  There is already
a use-case where the value of center can cause both to be true (if it is outside of the frame
boundary).  Seems like making that a hard error is unnecessary.