Transform data with ColorMapper (#4983)
* Have properties accept ColorMapper

* Refactor v_map_screen and use for v_compute too.

Part of v_map_screen was common in both LinearColorMapper and LogColorMapper
and was also duplicated.

Factor it out and add new _get_values method that gets the linear or log mapped
values. Use new _get_values method for v_compute that is needed for transform
to apply.

* Add an example very simply demonstrating the color mappers

* LinearColorMapper should just use palette length

* Revert "LinearColorMapper should just use palette length"

This reverts commit 5e04c102bead3a139d086a49c775c29d2d015a7c.

* Pass in a value for n

* Update texas example with ColorMapper

* Use ColorMapper for unemployment example

This saves ~30k (130k -> 101k) in the file size. In some use cases
the data savings are going to be remarkable.

* Add tests for compute method for data transform

They use the _get_values method of the color mappers.

* Check v_map_screen calls _get_values correctly

* Add more tests and update linear_color_mapper

* Update call to _get_values in v_compute

* Update example to include color bar.

* Simplified code does not require duplicate value at end of palette

Fix-up tests that check for duplicate value at the end

* Fix code quality & test description