Support multiple legend entries for same glyph renderer (#5089)
* Clarify draw_legend on glyph vs on glyph_renderer

Previously both glyph_renderer and glyph had the draw_legend function.
The glyph functions would "look up" a reference point - except that this
always returns none. Looking up a reference point is probably best done on
the glyph_renderer where actual data is stored.

This commit renames glyph.draw_legend to glyph.draw_legend_for_index and moves
the responsibility for looking up an index up to the glyph renderer.

In a future PR we will use this capability to allow drawing multiple legend entries for the
same glyph_renderer. For now, this should be a no-op change.

* get_reference_point can look up and index (w/tests)

Add capacity for get_reference_point to look-up an index based on a field
and value. This is currently not used as only call to get_reference_point
does not yet take advantage of this feature.

* Add label to Glyph

* Update daylight example to show new simplified legends adding.

* No-op cleaning-up of code in legend

* Support drawing legends from legend spec

Only using value at this point - updated daylight example now works.

* Support field as a label value

* Remove manual legend from gapminder

Use label property on glyph instead

* Add tests for Lgend.get_legend_names

* Add glyph renderer tests

* Code quality

* Add selenium tests for label powered legend

* Fix-up glyph tests

* Have figure use new legend options, update multiline example.

* Add tests for figure legend configuration

* Add example to docs

* Check for string_types not str