Return a factory function when window.document is not present
Previously jsdom was used to initialize window and document if
those weren't present, but this just complicates things, given
the multitude of environments bokehjs is supposed to work in.
Now, if window.document doesn't exist, a factory function is
returned instead of immediately initializing bokehjs. This
allows for proper configuration of the environment, without
any assumptions on bokehjs' side.