Fix off-by-one and missing test in LinearInterpolator (#5459)
There was no test that the final control point matched, which is how
this bug crept in. The problem was the predicate for .findLastIndex,
which was `x >= num`. This could return the last index for
@_x_sorted, and the subsequent `x2 = @_x_sorted[ind+1]` yielded null.

The predicate was changed to `num < x` which will always be at most the
next to last index. But this also necessitates special casing the first
control point, so the following test was added:
````
if x == @_x_sorted[0]
  return @_y_sorted[0]
```
which happens first.