Add size property to MultiSelect model
Add an attribute to toggle number of visible elements in MultiSelect.
Note that this can't exactly be controlled by CSS (the height property
can be used, but I'm not sure if there's an easy way to set it to
match the height of the options). Since the options can be dynamically
set, it seems reasonable to me to also let the size attribute be
dynamically set.