Set focus back to multiselect after on_change
Currently focus is lost from multiselect widget when a python
on_change callback is invoked. I'm not exactly clear on why that
happens but it seems to be sufficient to restore focus afterwards.
This makes it possible to scroll up/down in the select html element.

Also made a note in docs about browser quirks for size attr, and
updated example to use size attr.