Adding size attribute to MultiSelect model (#5580)
* Add size property to MultiSelect model

Add an attribute to toggle number of visible elements in MultiSelect.
Note that this can't exactly be controlled by CSS (the height property
can be used, but I'm not sure if there's an easy way to set it to
match the height of the options). Since the options can be dynamically
set, it seems reasonable to me to also let the size attribute be
dynamically set.

* Minor documentation adjustments.

Fixed a typo and made a brief note about BOKEH_MINIFIED env var,
in the case that we skip minifying during gulp build.

* Set focus back to multiselect after on_change

Currently focus is lost from multiselect widget when a python
on_change callback is invoked. I'm not exactly clear on why that
happens but it seems to be sufficient to restore focus afterwards.
This makes it possible to scroll up/down in the select html element.

Also made a note in docs about browser quirks for size attr, and
updated example to use size attr.