Allow to configure custom model base path and restore getcwd() (#5573)
Auto detection of base path is only possible if module of
a custom model's class is properly configured (has __file__).
That often doesn't happen, most notably in Jupyter Notebook,
so at least allow user to set a base path. getcwd() is restored,
because it's actually a good default in the Notebook.