Added auto conversion to ColumDataSource
Many times, we want to pass a data structure to a plotting function as
a data source. To do so, an explicit conversion is required. Often, the
converted object does not serve any purpose, except for the conversion
itself. This commit adds auto-conversion to reduce  boilerplate code.

Reference conversation: https://github.com/bokeh/bokeh/issues/5578