Columndatasource autoconvert (#5612)
* Added auto conversion to ColumDataSource

Many times, we want to pass a data structure to a plotting function as
a data source. To do so, an explicit conversion is required. Often, the
converted object does not serve any purpose, except for the conversion
itself. This commit adds auto-conversion to reduce  boilerplate code.

Reference conversation: https://github.com/bokeh/bokeh/issues/5578

* Implemented @bryevdv's PR comments

* Updating the documentation

As per @bryevdv request, updating the documentation in `figure.py`. The
only place that mentions `ColumnDataSource` is the docstring of
`scatter`.