Updating versioneer to 0.17+
Rebuilt versioneer using latest version, which has some windows
compatibility fixes and so forth. The main migration is moving
configuration for versioneer from setup.py to setup.cfg.

Refer to the python-versioneer repository for documentation;
roughly it installs a versioneer.py to the root directory of
the project which is used by setup.py to build a _version.py
file (a default _version.py is also provided when setup.py
is not used). The _version.py is used to retrieve the version
based on the VCS tag.