Updating versioneer to 0.17+ (#5622)
* Updating versioneer to 0.17+

Rebuilt versioneer using latest version, which has some windows
compatibility fixes and so forth. The main migration is moving
configuration for versioneer from setup.py to setup.cfg.

Refer to the python-versioneer repository for documentation;
roughly it installs a versioneer.py to the root directory of
the project which is used by setup.py to build a _version.py
file (a default _version.py is also provided when setup.py
is not used). The _version.py is used to retrieve the version
based on the VCS tag.

* Don't match tag_prefix in git describe

Not sure why but on windows, the asterisk is getting expanded
in subprocess.Popen(shell=False). Doing this as a temporary
workaround.

* Use style=git-describe for versioneer

To keep the __version__ string backwards-compatible, use the
git-describe style offered by versioneer. (By default, versioneer
uses a PEP440-compliant __version__ string which uses "+"
as a delimiter for local versioning information.)

* no empty lines at the top of source files

* Update _version.py

* Update _version.py

* remove trailing whitespace