Allow reselection of objects
Instead of computing the union of indices, simply concatenate the
previously selected indices with the currently provided ones. This
should allow for things like deselection on a second click. Though it
doesn't require that per se as it could also be interpreted as selecting
the same object again thus resulting in later selections being no-ops.