Remove ggplot.py examples and references.
Bokeh has not been able to be used with recent versions of ggplot.py
already for some time. The difficulty of installing older versions for
python 3.6 means the docs are effectively not buildable by python 3.6.

Given the questionable future for all MPL compat for Bokeh, this PR
takes the simple route for making the docs buildable across all Python
versions, and removes the handful of ggplot.py examples, which were not
very compelling to begin with.