Make sure properties are updated on (deferred) init (#5924)
Previously, a call to prop.update for all model properties happened in
attach_document. But attach_document was not being called for new models
on patches to existing documents, causing problems. This PR adds a
(deferred) initializer to HasProps that ensures proper property setup for
all models is completed in a consistent way.