Don't use "Dummy" bokeh models (#5972)
The original code created a CDS for every year of data, but only used
the first CDS in the document, switching out .data on upates. It's worth
investigating why this caused a problem separately, but this usage is
not advised in any case. (The new code is also faster) Bokeh models should
be created when they are intended to be put in a Document. Internal data
storage is best left to non-Bokeh data structures.