Make plots examples/charts/file/hover_span.py uniform #6093
The vline,hline,int_vline and int_hline plots in hover_span.py were
missing x='Year'.  This caused them to display different plots than the other
plots on the example.  Namely, the series did not overlap in the vertical.

When the series overlap in the vertical, vline hovertool causes hovers to overlap
each other, partially occluding the first drawn hover.

This example needs to be adjusted to demonstrate the problem with rendering multiple
hovertools.  Handling that issue will be done in another ticket.