Exclude `scripts` and `tests` in `setup`
Previously `setup` was including the `scripts` and `tests` in the PyPI
package and the `site-packages` install. Given both of these are not in
the Bokeh namespace, the `script` appear to be for deployment and CI
purposes only, and the `tests` have been intentionally moved out of the
Bokeh namespace, it makes to exclude these when looking for packages to
bundle and install with Bokeh. Hence we inform `setuptools` that it
should ignore these.