Include the tests in PyPI packages
When bundling up PyPI packages, make sure to include the `test`
directory and everything within it. Even though we do not want this
installed in `site-packages`, we still want access to this from any PyPI
deploys. This way a user can run the test suite to make sure their
install works if they need or a packager can verify that what they have
packaged functions correctly. Thus we include this directory in the
`MANIFEST.in` to ensure it ships with everything else.