Fix packaging (#6114)
* Exclude `scripts` and `tests` in `setup`

Previously `setup` was including the `scripts` and `tests` in the PyPI
package and the `site-packages` install. Given both of these are not in
the Bokeh namespace, the `script` appear to be for deployment and CI
purposes only, and the `tests` have been intentionally moved out of the
Bokeh namespace, it makes to exclude these when looking for packages to
bundle and install with Bokeh. Hence we inform `setuptools` that it
should ignore these.

* Include the tests in PyPI packages

When bundling up PyPI packages, make sure to include the `test`
directory and everything within it. Even though we do not want this
installed in `site-packages`, we still want access to this from any PyPI
deploys. This way a user can run the test suite to make sure their
install works if they need or a packager can verify that what they have
packaged functions correctly. Thus we include this directory in the
`MANIFEST.in` to ensure it ships with everything else.