Add hover tool support for Segment
issues: closes #6161

adds _hit_point and _hit_span to the Segment glyph so that it can work
with hover tools.

Also updates HoverTool docstring to indicate segments is not functional
with hover tools (as well as multi line which was erroneosly listed as
unsupported)

Also makes a minor change to line hit-testing. From inspection the
existing code would appear to fail on lines that "go backwards"